docs: replace bug reports with fix requests
