+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -51,5 +51,5 @@ body:
     attributes:
       label: Before submitting
       options:
-        - label: I checked existing issues and pull requests.
+        - label: I checked existing requests and pull requests.
           required: true
