+++ docs/content/architecture/overview.md
@@ -149,7 +149,7 @@ Vize crates are named after **art and sculpture terminology**, reflecting how ea
 | **Atelier**  | /ˌætəlˈjeɪ/  | Artist's workshop where creation happens                 | Compiler workspaces — where code is transformed into its final form            |
 | **Vitrine**  | /vɪˈtriːn/   | Glass display case in a museum                           | Bindings — a transparent layer that exposes the compiler to external consumers |
 | **Canon**    | /ˈkænən/     | Standard of ideal proportions in classical sculpture     | Type checker — ensures code conforms to the standard of correctness            |
-| **Patina**   | /ˈpætɪnə/    | Aged surface finish that indicates quality and care      | Linter — polishes code by identifying issues that affect quality               |
+| **Patina**   | /ˈpætɪnə/    | Aged surface finish that indicates quality and care      | Linter — polishes code by identifying problems that affect quality             |
 | **Glyph**    | /ɡlɪf/       | Carved symbol or letterform with precise proportions     | Formatter — shapes code into consistent, readable letterforms                  |
 | **Maestro**  | /ˈmaɪstroʊ/  | Master conductor who orchestrates an ensemble            | LSP — orchestrates all language features into a unified editor experience      |
 | **Musea**    | /mjuːˈziːə/  | Plural of museum — a space for exhibiting art            | Component gallery — a space for exhibiting and exploring components            |
